Conduit.jl

Conduit.CDatatype — Type

CDatatype::Type

Conduit.CNode — Type

CNode::Type

Conduit.blueprint_mcarray_examples_xyz — Method

blueprint_mcarray_examples_xyz(mcarray_type::AbstractString, npts::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mcarray_examples_xyz — Method

blueprint_mcarray_examples_xyz(mcarray_type::AbstractString, npts::Integer, res::Node)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mcarray_is_interleaved — Method

blueprint_mcarray_is_interleaved(cnode::CNode)::Bool

Conduit.blueprint_mcarray_is_interleaved — Method

blueprint_mcarray_is_interleaved(node::Node)::Bool

Conduit.blueprint_mcarray_to_contiguous — Method

blueprint_mcarray_to_contiguous(cnode::CNode, cdest::CNode)::Bool

Result: cdest.

Conduit.blueprint_mcarray_to_contiguous — Method

blueprint_mcarray_to_contiguous(node::Node, dest::Node)::Bool

Result: dest.

Conduit.blueprint_mcarray_to_interleaved — Method

blueprint_mcarray_to_interleaved(cnode::CNode, cdest::CNode)::Bool

Result: cdest.

Conduit.blueprint_mcarray_to_interleaved — Method

blueprint_mcarray_to_interleaved(node::Node, dest::Node)::Bool

Result: dest.

Conduit.blueprint_mcarray_verify — Method

blueprint_mcarray_verify(cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node confirms to the blueprint mcarray protocol.

Conduit.blueprint_mcarray_verify — Method

blueprint_mcarray_verify(node::Node, info::Node)::Bool

Result: info.

Verify passed node confirms to the blueprint mcarray protocol.

Conduit.blueprint_mcarray_verify_sub_protocol — Method

blueprint_mcarray_verify_sub_protocol(protocol::AbstractString, cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node confirms to given blueprint mcarray sub protocol.

Conduit.blueprint_mcarray_verify_sub_protocol — Method

blueprint_mcarray_verify_sub_protocol(protocol::AbstractString, node::Node, info::Node)::Bool

Result: info.

Verify passed node confirms to given blueprint mcarray sub protocol.

Conduit.blueprint_mesh_examples_basic — Method

blueprint_mesh_examples_basic(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_basic — Method

blueprint_mesh_examples_basic(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, res::Node)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_braid — Method

blueprint_mesh_examples_braid(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_braid — Method

blueprint_mesh_examples_braid(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, res::Node)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_generate — Method

blueprint_mesh_examples_generate(example_name::AbstractString, copts::CNode, cres::CNode)

Result: cres.

generates a named example mesh using options. expects opts node to contains relevant args for each example.

Conduit.blueprint_mesh_examples_generate_default_options — Method

blueprint_mesh_examples_generate_default_options(example_name::AbstractString, copts::CNode)

Result: copts.

creates the default options for a given example

Conduit.blueprint_mesh_examples_julia — Method

blueprint_mesh_examples_julia(
    nx::Integer,
    ny::Integer,
    x_min::AbstractFloat,
    x_max::AbstractFloat,
    y_min::AbstractFloat,
    y_max::AbstractFloat,
    c_re::AbstractFloat,
    c_im::AbstractFloat,
    cres::CNode,
)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_julia — Method

blueprint_mesh_examples_julia(
    nx::Integer,
    ny::Integer,
    x_min::AbstractFloat,
    x_max::AbstractFloat,
    y_min::AbstractFloat,
    y_max::AbstractFloat,
    c_re::AbstractFloat,
    c_im::AbstractFloat,
    res::Node,
)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_misc — Method

blueprint_mesh_examples_misc(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_misc — Method

blueprint_mesh_examples_misc(mesh_type::AbstractString, nx::Integer, ny::Integer, nz::Integer, res::Node)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_polytess — Method

blueprint_mesh_examples_polytess(nlevels::Integer, nz::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_polytess — Method

blueprint_mesh_examples_polytess(nlevels::Integer, nz::Integer, res::Node)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_spiral — Method

blueprint_mesh_examples_spiral(ndomains::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_examples_spiral — Method

blueprint_mesh_examples_spiral(ndomains::Integer, res::Node)::Nothing

Result: res.

Interface to generate example mesh blueprint data.

Conduit.blueprint_mesh_flatten — Method

blueprint_mesh_flatten(cmesh::CNode, coptions::CNode, coutput::CNode)::Nothing

Result: coutput.

Flatten a mesh to a table

Conduit.blueprint_mesh_flatten — Method

blueprint_mesh_flatten(mesh::Node, options::Node, output::Node)::Nothing

Result: output.

Flatten a mesh to a table

Conduit.blueprint_mesh_generate_index — Method

blueprint_mesh_generate_index(cmesh::CNode, ref_path::AbstractString, num_domains::Integer, cindex_out::CNode)::Nothing

Result: cindex_out.

Generate mesh::index from valid mesh.

Conduit.blueprint_mesh_generate_index — Method

blueprint_mesh_generate_index(mesh::Node, ref_path::AbstractString, num_domains::Integer, index_out::Node)::Nothing

Result: index_out.

Generate mesh::index from valid mesh.

Conduit.blueprint_mesh_partition — Method

blueprint_mesh_partition(cmesh::CNode, coptions::CNode, coutput::CNode)::Nothing

Result: coutput.

Partition a mesh

Conduit.blueprint_mesh_partition — Method

blueprint_mesh_partition(mesh::Node, options::Node, output::Node)::Nothing

Result: output.

Partition a mesh

Conduit.blueprint_mesh_verify — Method

blueprint_mesh_verify(cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node confirms to the blueprint mesh protocol.

Conduit.blueprint_mesh_verify — Method

blueprint_mesh_verify(node::Node, info::Node)::Bool

Result: info.

Verify passed node confirms to the blueprint mesh protocol.

Conduit.blueprint_mesh_verify_sub_protocol — Method

blueprint_mesh_verify_sub_protocol(protocol::AbstractString, cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node confirms to given blueprint mesh sub protocol.

Conduit.blueprint_mesh_verify_sub_protocol — Method

blueprint_mesh_verify_sub_protocol(protocol::AbstractString, node::Node, info::Node)::Bool

Result: info.

Verify passed node confirms to given blueprint mesh sub protocol.

Conduit.blueprint_table_examples_basic — Method

blueprint_table_examples_basic(nx::Integer, ny::Integer, nz::Integer, cres::CNode)::Nothing

Result: cres.

Interface to generate example table blueprint data.

Conduit.blueprint_table_examples_basic — Method

blueprint_table_examples_basic(nx::Integer, ny::Integer, nz::Integer, res::Node)::Nothing

Result: res.

Interface to generate example table blueprint data.

Conduit.blueprint_table_verify — Method

blueprint_table_verify(cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node conforms to the blueprint table protocol.

Conduit.blueprint_table_verify — Method

blueprint_table_verify(node::Node, info::Node)::Bool

Result: info.

Verify passed node conforms to the blueprint table protocol.

Conduit.blueprint_table_verify_sub_protocol — Method

blueprint_table_verify_sub_protocol(protocol::AbstractString, cnode::CNode, cinfo::CNode)::Bool

Result: cinfo.

Verify passed node conforms to the blueprint table protocol.

Conduit.blueprint_table_verify_sub_protocol — Method

blueprint_table_verify_sub_protocol(protocol::AbstractString, node::Node, info::Node)::Bool

Result: info.

Verify passed node conforms to the blueprint table protocol.